refactor(engine): replace three PR-recovery paths with single owner
Deletes the three divergent "PR terminal → advance" code paths and
replaces them with a single call to runValidatePRTerminalAdvance:

1. Phase-2 Validate merged-PR inline fallback (poll.go: after
   attemptMergeOnValidate fails, the inner PR-fetch-and-advance block).
2. runPausedItemMergedPRRecovery and its call site (paused_recovery.go
   deleted entirely; paused_recovery_test.go deleted with it).
3. Convergence-paused recovery loop (poll.go: the loop over paused items
   without fabrik:awaiting-ci).

The deleted paths used hand-maintained label negations to carve up the
item space; any item that fell through the intersection was stranded
(see issue #874). The single owner advances Validate-stage items
unconditionally on a merged PR regardless of gate label.

Tests for the deleted code are replaced by the table-driven tests and
regression test in pr_terminal_advance_test.go (committed previously).
